feat: refactor screenshot logic
